Ensure deleting-compaction-strategy removes row tombstones
When a full row is deleted, Cassandra creates a row tombstone, also
known as a top-level deletion, which is different from both a range
tombstone and a regular tombstone. This tombstone is handled specially
inside Cassandra, and never yielded by any OnDiskAtomIterator. Instead,
it is tucked away inside the metadata inside the column family accessor
of the OnDiskAtomIterator.

This means that a row tombstone will never be deleted by the deleting
compaction strategy, unless the whole row is deleted. Concretely, this
means that when using the RuleBasedLateTTLConvictor on a table where
full rows have been deleted, the row tombstones will not be deleted.

This introduces a new method to the IOnDiskAtomFilter to allow deletion
of row tombstones and uses that in the RuleBasedLateTTLConvictor to
signal that tombstone within the partition TTL should be removed.