Only check top-level deletion if it's a tombstone
Rows without a top-level tombstone are represented using a "live"
deletion time, which effectively means that the row hasn't been deleted.
In principle, we could determine if we should keep this tombstone, but
given that the only option we give the atom filter is whether it should
be included or not, and thus the effect of both branches should be that
we get a "live" top-level deletion. In that case, it seems better to
avoid all the hoops and allocations, and just return the original as is.